Commit Eventos

+++ app-modules/events/database/migrations/2025_07_08_211648_add_column_table_event.php
@@ -6,23 +6,21 @@ use Illuminate\Support\Facades\Schema;
 
 return new class extends Migration
 {
-    /**
-     * Run the migrations.
-     */
     public function up(): void
     {
-        Schema::table('events', function (Blueprint $table) {
-            $table->string('ruta_img')->default('');
-        });
+        if (Schema::hasTable('events') && ! Schema::hasColumn('events', 'ruta_img')) {
+            Schema::table('events', function (Blueprint $table) {
+                $table->string('ruta_img')->default('');
+            });
+        }
     }
 
-    /**
-     * Reverse the migrations.
-     */
     public function down(): void
     {
-        Schema::table('events', function (Blueprint $table) {
-            $table->dropColumn('ruta_img');
-        });
+        if (Schema::hasTable('events') && Schema::hasColumn('events', 'ruta_img')) {
+            Schema::table('events', function (Blueprint $table) {
+                $table->dropColumn('ruta_img');
+            });
+        }
     }
 };
