+++ app-modules/events/database/migrations/2025_08_12_232217_remove_column_ruta_img_from_events_table.php
@@ -11,9 +11,11 @@ return new class extends Migration
      */
     public function up(): void
     {
-        Schema::table('events', function (Blueprint $table) {
-            $table->dropColumn('ruta_img');
-        });
+        if (Schema::hasTable('events') && Schema::hasColumn('events', 'ruta_img')) {
+            Schema::table('events', function (Blueprint $table) {
+                $table->dropColumn('ruta_img');
+            });
+        }
     }
 
     /**
@@ -21,8 +23,10 @@ return new class extends Migration
      */
     public function down(): void
     {
-        Schema::table('events', function (Blueprint $table) {
-            $table->string('ruta_img')->nullable();
-        });
+        if (Schema::hasTable('events') && ! Schema::hasColumn('events', 'ruta_img')) {
+            Schema::table('events', function (Blueprint $table) {
+                $table->string('ruta_img')->nullable();
+            });
+        }
     }
 };
